Updated endpoints and relations

+++ Farm/src/main/java/at/htl/farm/rest/FarmEndpoint.java
@@ -1,5 +1,6 @@
 package at.htl.farm.rest;
 
+import at.htl.farm.model.Animal;
 import at.htl.farm.model.Farm;
 
 import javax.ejb.Stateless;
@@ -7,7 +8,9 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 import javax.ws.rs.*;
+import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Path("farm")
@@ -19,38 +22,42 @@ public class FarmEndpoint {
 
     @Path("findall")
     @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public List<Farm> findAll(){
-        return em.createNamedQuery("Farm.findAll", Farm.class).getResultList();
+    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+    public Response findAll() {
+        List<Farm> farmList = em.createNamedQuery("Farm.findAll", Farm.class).getResultList();
+        GenericEntity entity = new GenericEntity<List<Farm>>(farmList){};
+
+        if(farmList != null && !farmList.isEmpty())
+            return Response.ok(entity).build();
+        else
+            return Response.noContent().build();
+    }
+
+    @Path("find/{id}")
+    @GET
+    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+    public Response find(@PathParam("id") long id) {
+        Farm farm = em.createNamedQuery("Farm.findById", Farm.class).setParameter(1, id).getSingleResult();
+
+        if(farm != null) {
+            return Response.ok(farm).build();
+        } else {
+            return Response.noContent().build();
+        }
     }
 
     @Path("delete/{id}")
     @DELETE
-    @Transactional
-    public void delete(@PathParam("id") long id){
-        Farm f = em.find(Farm.class, id);
-        em.remove(f);
-    }
-
-    @Path("put/{id}")
-    @PUT
-    @Transactional
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void put(@PathParam("id") long id, Farm farm){
-        Farm f = em.find(Farm.class, id);
-        f.setLocation(farm.getLocation());
-        em.merge(f);
-    }
-
-    @Path("post")
-    @POST
-    @Transactional
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.TEXT_PLAIN)
-    public Long post(Farm farm){
-        em.persist(farm);
-        em.flush();
-        return farm.getId();
+    public Response delete(@PathParam("id") long id) {
+        try{
+            Farm f = em.find(Farm.class, id);
+            if(f != null){
+                em.remove(f);
+            }
+        }catch (Exception e){
+            return Response.status(404).build();
+        }
+        return Response.ok().build();
     }
 
 }
