+++ Farm/src/main/java/at/htl/farm/rest/FieldEndpoint.java
@@ -5,9 +5,11 @@ import at.htl.farm.model.Field;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.transaction.Transactional;
 import javax.ws.rs.*;
+import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import java.net.URI;
 import java.util.List;
 
 @Path("field")
@@ -19,53 +21,66 @@ public class FieldEndpoint {
 
     @Path("findall")
     @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public List<Field> findAll(){
-        return em.createNamedQuery("Field.findAll", Field.class).getResultList();
+    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+    public Response findAll() {
+        List<Field> fieldList = em.createNamedQuery("Field.findAll", Field.class).getResultList();
+        GenericEntity entity = new GenericEntity<List<Field>>(fieldList){};
+
+        if (fieldList != null && !fieldList.isEmpty()) {
+            return Response.ok(entity).build();
+        } else {
+            return Response.noContent().build();
+        }
     }
 
-    @Path("findById/{id}")
+    @Path("find/{id}")
     @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public Field findById(@PathParam("id") long id){
-        return em.createNamedQuery("Field.findById", Field.class).setParameter(1, id).getSingleResult();
-    }
+    @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+    public Response findById(@PathParam("id") long id) {
+        Field f = em.find(Field.class, id);
 
-    @Path("findBySeed/{seed}")
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public Field findBySeed(@PathParam("seed") String seed){
-        return em.createNamedQuery("Field.findBySeed", Field.class).setParameter(1, seed).getSingleResult();
+        if (f != null) {
+            return Response.ok(f).build();
+        } else {
+            return Response.noContent().build();
+        }
     }
 
     @Path("delete/{id}")
     @DELETE
-    @Transactional
-    public void delete(@PathParam("id") long id){
-        Field f = em.find(Field.class, id);
-        em.remove(f);
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response delete(@PathParam("id") long id) {
+        try {
+            Field f = em.find(Field.class, id);
+            if (f != null) {
+                //f.getFarm().removeField(f);
+                em.remove(f);
+            }
+        } catch (Exception ex) {
+            return Response.status(404).build();
+        }
+
+        return Response.ok().build();
     }
 
-    @Path("put/{id}")
-    @PUT
-    @Transactional
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void put(@PathParam("id") long id, Field field){
-        Field f = em.find(Field.class, id);
-        f.setFarm(field.getFarm());
-        f.setHectare(field.getHectare());
-        f.setPlantedSeeds(field.getPlantedSeeds());
-        em.merge(f);
-    }
-
-    @Path("post")
+    @Path("new")
     @POST
-    @Transactional
     @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.TEXT_PLAIN)
-    public Long post(Field field){
-        em.persist(field);
+    public Response newField(Field f) {
+        em.persist(f);
         em.flush();
-        return field.getId();
+        return Response.created(URI.create("http://localhost:8080/farm/rs/field/find/" + f.getId())).build();
+    }
+
+    @Path("update/{id}")
+    @PUT
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response update(@PathParam("id") long id, Field f){
+        Field field = em.find(Field.class, id);
+        field.setFarm(f.getFarm());
+        field.setHectare(f.getHectare());
+        field.setPlantedSeeds(f.getPlantedSeeds());
+        em.merge(field);
+        return Response.ok().build();
     }
 }
